chore: move SyncTokenAwareClient to the common test helpers
